# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Bill:
    properties:
      bill_category:
        $ref: '#/definitions/models.BillCategory'
      bill_category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      id:
        type: integer
      is_payed:
        type: boolean
      ocr_data:
        items:
          type: integer
        type: array
      payment_date:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      public:
        description: separates home/private budget
        type: boolean
      receipt_image:
        type: string
      splits:
        items:
          $ref: '#/definitions/models.BillSplit'
        type: array
      total_amount:
        type: number
      type:
        description: Kept for backward compatibility or as fallback
        type: string
      uploaded_by:
        type: integer
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.BillCategory:
    properties:
      color:
        description: Hex color
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      public:
        type: boolean
    type: object
  models.BillListResponse:
    properties:
      bills:
        items:
          $ref: '#/definitions/models.Bill'
        type: array
      status:
        example: true
        type: boolean
    type: object
  models.BillSplit:
    properties:
      amount:
        type: number
      bill:
        $ref: '#/definitions/models.Bill'
      bill_id:
        type: integer
      id:
        type: integer
      paid:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
...
  models.CreateBillRequest:
    properties:
      bill_category_id:
        type: integer
      description:
        type: string
      is_regular:
        type: boolean
      ocr_data:
        items:
          type: integer
        type: array
      period_end:
        type: string
      period_start:
        type: string
      public:
        type: boolean
      receipt_image:
        type: string
      recurrence_day:
        type: integer
      recurrence_type:
        type: string
      splits:
        items:
          $ref: '#/definitions/models.SplitInput'
        type: array
      total_amount:
        minimum: 0
        type: number
      type:
        description: Optional if CategoryID is provided
        type: string
    required:
    - ocr_data
    - period_end
    - period_start
    - total_amount
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: invalid home ID
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  models.Home:
    properties:
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      invite_code:
        type: string
      memberships:
        description: relations
        items:
          $ref: '#/definitions/models.HomeMembership'
        type: array
      name:
        type: string
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
  models.HomeMembership:
    properties:
      created_at:
        type: string
      home:
        $ref: '#/definitions/models.Home'
      home_id:
        type: integer
      id:
        type: integer
      joined_at:
        type: string
      role:
        type: string
      status:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: relations
      user_id:
        type: integer
    type: object
...
  models.Room:
    properties:
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
    type: object
...
  models.SplitInput:
    properties:
      amount:
        type: number
      user_id:
        type: integer
    type: object
  models.Task:
    properties:
      assignments:
        items:
          $ref: '#/definitions/models.TaskAssignment'
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/models.User'
      description:
        type: string
      due_date:
        type: string
      home:
        allOf:
        - $ref: '#/definitions/models.Home'
        description: relations
      home_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      reminder_minutes:
        type: integer
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: integer
      schedule:
        $ref: '#/definitions/models.TaskSchedule'
      schedule_type:
        type: string
    type: object
  models.TaskAssignment:
    properties:
      assigned_date:
        type: string
      complete_date:
        type: string
      created_at:
        type: string
      id:
        type: integer
      reminder_sent_at:
        type: string
      status:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: relations
      task_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.TaskSchedule:
    properties:
      created_at:
        type: string
      current_rotation_index:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      next_run_date:
        type: string
      recurrence_type:
        description: daily, weekly, monthly
        type: string
      rotation_user_ids:
        description: JSON array of user IDs in order
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: relations
      task_id:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      memberships:
        description: relations
        items:
          $ref: '#/definitions/models.HomeMembership'
        type: array
      name:
        type: string
      profile_public:
        type: boolean
      task_assignments:
        items:
          $ref: '#/definitions/models.TaskAssignment'
        type: array
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Diploma project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Diploma Server API
  version: "1.0"
paths:
...
  /homes/{home_id}/bills:
    get:
      description: Get all bills in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get bills by home ID
      tags:
      - bill
    post:
      consumes:
      - application/json
      description: Create a new bill in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Create Bill Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateBillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new bill
      tags:
      - bill
  /homes/{home_id}/bills/{bill_id}:
    delete:
      description: Delete a bill by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Bill ID
        in: path
        name: bill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete bill
      tags:
      - bill
    get:
      description: Get bill details by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Bill ID
        in: path
        name: bill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get bill by ID
      tags:
      - bill
    patch:
      description: Mark a bill as payed
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Bill ID
        in: path
        name: bill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark bill as payed
      tags:
      - bill
  /homes/{home_id}/bills/{bill_id}/splits:
    put:
      consumes:
      - application/json
      description: Update how a bill is split between users (uploader or admin only)
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Bill ID
        in: path
        name: bill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update bill splits
      tags:
      - bill
  /homes/{home_id}/bills/{bill_id}/splits/{split_id}/paid:
    patch:
      description: Mark a single user's bill split as paid
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Bill ID
        in: path
        name: bill_id
        required: true
        type: integer
      - description: Split ID
        in: path
        name: split_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark a split as paid
      tags:
      - bill
  /homes/{home_id}/bills/private:
    get:
      description: Returns bills in the selected home where public=false and uploaded_by
        is the authenticated user. The user must be a member of the home.
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Filter by bill category ID
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Private bills list
          schema:
            $ref: '#/definitions/models.BillListResponse'
        "400":
          description: Invalid home_id or category_id
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Missing or invalid authentication
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: User is not a member of the home
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to retrieve bills
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List current user's private bills
      tags:
      - bill
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"